+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### 2.6.8 - 2024-08-28
+* 274258: Add instructions link to settings
 * 274259: Display error message when payment methods are not available
 
 ### 2.6.7 - 2024-06-25